fix(msg-nostr-read): Cleaned up code
